Struct wai_bindgen_gen_wasmer_py::dependencies::Dependencies

pub struct Dependencies {

    pub needs_clamp: bool,
    pub needs_store: bool,
    pub needs_load: bool,
    pub needs_validate_guest_char: bool,
    pub needs_expected: bool,
    pub needs_i32_to_f32: bool,
    pub needs_f32_to_i32: bool,
    pub needs_i64_to_f64: bool,
    pub needs_f64_to_i64: bool,
    pub needs_decode_utf8: bool,
    pub needs_encode_utf8: bool,
    pub needs_list_canon_lift: bool,
    pub needs_list_canon_lower: bool,
    pub needs_t_typevar: bool,
    pub needs_resources: bool,
    pub pyimports: BTreeMap<String, Option<BTreeSet<String>>>,

}

Tracks all of the import and intrinsics that a given codegen requires and how to generate them when needed.

Fields
needs_clamp: boolneeds_store: boolneeds_load: boolneeds_validate_guest_char: boolneeds_expected: boolneeds_i32_to_f32: boolneeds_f32_to_i32: boolneeds_i64_to_f64: boolneeds_f64_to_i64: boolneeds_decode_utf8: boolneeds_encode_utf8: boolneeds_list_canon_lift: boolneeds_list_canon_lower: boolneeds_t_typevar: boolneeds_resources: boolpyimports: BTreeMap<String, Option<BTreeSet<String>>>
Implementations

impl Dependencies

pub fn pyimport<'a>(&mut self, module: &str, name: impl Into<Option<&'a str>>)

Record that a Python import is required

Examples

// Import a specific item from a module
deps.pyimport("typing", "NamedTuple");
// Import an entire module
deps.pyimport("collections", None);

pub fn intrinsics(&mut self) -> Source

Create a Source containing all of the intrinsics required according to this Dependencies struct.